+++ client/src/index.tsx
@@ -1,22 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
+import {BrowserRouter} from 'react-router-dom'
 import App from './App';
-import StoreList from './StoreList';
 import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(
-  <React.StrictMode>
+  <BrowserRouter>
     <App />
-  </React.StrictMode>,
-  document.getElementById('header')
-);
-
-ReactDOM.render(
-  <React.StrictMode>
-    <StoreList />
-  </React.StrictMode>,
-  document.getElementById('stores')
+  </BrowserRouter>,
+  document.getElementById('root')
 );
 
 // If you want your app to work offline and load faster, you can change
