+++ client/src/pages/NotFound.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function NotFound() {
+  return (
+    <div className="NotFound">
+      <h1> 404 - Page Not Found </h1>
+    </div>
+  )
+}
+
+export default NotFound
